[#59645746] Style Tiles (Guide)
This style guide will set the tone for all views. If / when later on in
the project a new view is needed to be created you can reuse
predesigned elements to achieve the desired look for the view without
having to go through the wireframe process, speeding your development
time.

Please use the resources linked below for further explanation. These
assets you create should be placed into an ASSETS folder in your Github
Repo.

http://styletil.es/